Feature/change ticket type

+++ data/ticket/src/main/java/jp/kztproject/rewardedtodo/data/ticket/ITicketRepository.kt
@@ -3,10 +3,10 @@ package jp.kztproject.rewardedtodo.data.ticket
 import kotlinx.coroutines.flow.Flow
 
 interface ITicketRepository {
-    suspend fun addTicket(numberOfTicket: Float)
+    suspend fun addTicket(numberOfTicket: Int)
 
     suspend fun consumeTicket()
 
     // TODO change return type to domain model
-    suspend fun getNumberOfTicket(): Flow<Float>
+    suspend fun getNumberOfTicket(): Flow<Int>
 }
