+++ feature/todo/src/main/java/jp/kztproject/rewardedtodo/todo/domain/Todo.kt
@@ -4,11 +4,11 @@ data class Todo(
         val id: Long?,
         val todoistId: Long?,
         var name: String,   //TODO divide model
-        val numberOfTicketsObtained: Float,
+        val numberOfTicketsObtained: Int,
         val isRepeat: Boolean
 ) {
     companion object {
-        const val DEFAULT_NUMBER_OF_TICHKET = 1.0f
+        const val DEFAULT_NUMBER_OF_TICHKET = 1
     }
     fun hasTodoistId(): Boolean {
         return todoistId != null
